+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Yii2 Queue Extension Change Log
 -----------------------
 
 - Enh #430: Added configurable AMQP Exchange type (s1lver)
+- Enh #435: Added the ability to set optional arguments for the AMQP queue (s1lver)
 
 
 2.3.2 May 05, 2021